// src/frontend/src/backend.d.ts
export type Time = bigint;
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface Repeater {
    id: bigint;
    status: Status;
    dcsCode: string;
    coverageDescription: string;
    city: string;
    toneMode: string;
    submittedBy: string;
    offset: number;
    submissionStatus: SubmissionStatus;
    callSign: string;
    zipCode: string;
    state: string;
    timestamp: Time;
    sponsor: string;
    autopatchInfo: string;
    frequency: number;
    operationalNotes: string;
    linkInfo: string;
    ctcssTone: string;
}
// ...
export enum Status {
    active = "active",
    inactive = "inactive"
}
export enum SubmissionStatus {
    pending = "pending",
    approved = "approved",
    rejected = "rejected"
}